requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
flask-sock==0.5.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.1
pytz==2022.1
requests==2.27.1
simple-websocket==0.5.2
SQLAlchemy==1.4.37
Turbo-Flask==0.8.0
urllib3==1.26.9
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.8.0
